requirements.txt
Node
streamlit
pypdf2
faiss-cpu
tiktoken
openai
pandas
python-dotenv
pdfkit
pyppeteer
beautifulsoup4 
